LayoutTests/fast/shadow-dom/shadow-at-root-during-disconnect-expected.txt

ALERT: false
ALERT: true
Tests proper tear-down of shadow content.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Passes if the second alert is 'true'.
PASS successfullyParsed is true

TEST COMPLETE
